+++ packages/cli/src/controller/init-controller.ts
@@ -1,30 +1,98 @@
-// Copyright 2020-2021 OnFinality Limited authors & contributors
+// Copyright 2020-2022 OnFinality Limited authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import childProcess from 'child_process';
+import childProcess, {execSync} from 'child_process';
 import fs from 'fs';
 import * as path from 'path';
 import {promisify} from 'util';
 import {ProjectManifestV0_0_1, ProjectManifestV0_2_0} from '@subql/common';
+import axios from 'axios';
 import yaml from 'js-yaml';
 import rimraf from 'rimraf';
-import simpleGit from 'simple-git';
+import git from 'simple-git';
 import {isProjectSpecV0_2_0, ProjectSpecBase} from '../types';
 
-const STARTER_PATH = 'https://github.com/subquery/subql-starter';
+const TEMPLATES_REMOTE = 'https://raw.githubusercontent.com/subquery/templates/main/templates.json';
 
-export async function createProject(localPath: string, project: ProjectSpecBase): Promise<string> {
-  const projectPath = path.join(localPath, project.name);
+export interface Template {
+  name: string;
+  description: string;
+  remote: string;
+  branch: string;
+  network: string;
+  specVersion: string;
+}
 
-  const cloneArgs = isProjectSpecV0_2_0(project)
-    ? ['-b', 'v0.2.0', '--single-branch']
-    : ['-b', 'v0.0.1', '--single-branch'];
+export async function fetchTemplates(remote: string = TEMPLATES_REMOTE): Promise<Template[]> {
+  return axios
+    .create()
+    .get(remote)
+    .then(({data}) => data as Template[])
+    .catch((err) => {
+      throw new Error(`Unable to reach endpoint '${remote}', ${err}`);
+    });
+}
 
+export async function cloneProjectGit(
+  localPath: string,
+  projectName: string,
+  projectRemote: string,
+  branch: string
+): Promise<string> {
+  const projectPath = path.join(localPath, projectName);
   try {
-    await simpleGit().clone(STARTER_PATH, projectPath, cloneArgs);
+    await git().clone(projectRemote, projectPath, ['-b', branch, '--single-branch']);
   } catch (e) {
-    throw new Error('Failed to clone starter template from git');
+    let err = 'Failed to clone starter template from git';
+    try {
+      execSync('git --version');
+    } catch (_) {
+      err += ', please install git and ensure that it is available from command line';
+    }
+    throw new Error(err);
   }
+  return projectPath;
+}
+
+export async function cloneProjectTemplate(
+  localPath: string,
+  projectName: string,
+  template: Template
+): Promise<string> {
+  const projectPath = path.join(localPath, projectName);
+  try {
+    await git().clone(template.remote, projectPath, ['-b', template.branch, '--single-branch']);
+  } catch (e) {
+    let err = 'Failed to clone starter template from git';
+    try {
+      execSync('git --version');
+    } catch (_) {
+      err += ', please install git and ensure that it is available from command line';
+    }
+    throw new Error(err);
+  }
+  return projectPath;
+}
+
+export async function readDefaults(projectPath: string): Promise<string[]> {
+  const packageData = await fs.promises.readFile(`${projectPath}/package.json`);
+  const currentPackage = JSON.parse(packageData.toString());
+
+  const yamlPath = path.join(`${projectPath}`, `project.yaml`);
+  const manifest = await fs.promises.readFile(yamlPath, 'utf8');
+  const currentProject = yaml.load(manifest) as ProjectManifestV0_0_1 | ProjectManifestV0_2_0;
+  return [
+    currentProject.specVersion,
+    currentProject.repository,
+    currentProject.network.endpoint,
+    currentPackage.author,
+    currentPackage.version,
+    currentPackage.description,
+    currentPackage.license,
+  ];
+}
+
+export async function prepare(projectPath: string, project: ProjectSpecBase): Promise<void> {
   try {
     await prepareManifest(projectPath, project);
   } catch (e) {
@@ -38,10 +106,13 @@ export async function createProject(localPath: string, project: ProjectSpecBase)
   try {
     await promisify(rimraf)(`${projectPath}/.git`);
   } catch (e) {
-    throw new Error('Failed to remove .git from starter project');
+    throw new Error('Failed to remove .git from template project');
   }
-
-  return projectPath;
+  try {
+    await promisify(rimraf)(`${projectPath}/.github`);
+  } catch (e) {
+    throw new Error('Failed to remove .github from template project');
+  }
 }
 
 async function preparePackage(projectPath: string, project: ProjectSpecBase): Promise<void> {