+++ packages/cli/src/controller/migrate-controller.ts
@@ -1,14 +1,14 @@
-// Copyright 2020-2021 OnFinality Limited authors & contributors
+// Copyright 2020-2022 OnFinality Limited authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
 import fs from 'fs';
 import path from 'path';
 import {
-  loadProjectManifest,
   ProjectManifestV0_2_0,
   ProjectManifestVersioned,
   ProjectNetworkV0_0_1,
   ChainTypes,
+  loadProjectManifest,
 } from '@subql/common';
 import {cli} from 'cli-ux';
 import yaml from 'js-yaml';