+++ packages/common/src/project/versioned/ProjectManifestVersioned.ts
@@ -1,18 +1,19 @@
-// Copyright 2020-2021 OnFinality Limited authors & contributors
+// Copyright 2020-2022 OnFinality Limited authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
 import {SubqlDatasource} from '@subql/types';
 import {plainToClass} from 'class-transformer';
-import {validateSync} from 'class-validator';
 import {IProjectManifest} from '../types';
 import {ProjectManifestV0_0_1Impl} from './v0_0_1';
 import {ProjectManifestV0_2_0Impl} from './v0_2_0';
+import {ProjectManifestV0_2_1Impl} from './v0_2_1';
 
 export type VersionedProjectManifest = {specVersion: string};
 
 const SUPPORTED_VERSIONS = {
   '0.0.1': ProjectManifestV0_0_1Impl,
   '0.2.0': ProjectManifestV0_2_0Impl,
+  '0.2.1': ProjectManifestV0_2_1Impl,
 };
 
 type Versions = keyof typeof SUPPORTED_VERSIONS;
@@ -25,6 +26,10 @@ export function manifestIsV0_0_1(manifest: IProjectManifest): manifest is Projec
 
 export function manifestIsV0_2_0(manifest: IProjectManifest): manifest is ProjectManifestV0_2_0Impl {
   return manifest.specVersion === '0.2.0';
+}
+
+export function manifestIsV0_2_1(manifest: IProjectManifest): manifest is ProjectManifestV0_2_1Impl {
+  return manifest.specVersion === '0.2.1';
 }
 
 export class ProjectManifestVersioned implements IProjectManifest {
@@ -54,8 +59,16 @@ export class ProjectManifestVersioned implements IProjectManifest {
     return this.specVersion === '0.2.0';
   }
 
+  get isV0_2_1(): boolean {
+    return this.specVersion === '0.2.1';
+  }
+
   get asV0_2_0(): ProjectManifestV0_2_0Impl {
     return this._impl as ProjectManifestV0_2_0Impl;
+  }
+
+  get asV0_2_1(): ProjectManifestV0_2_1Impl {
+    return this._impl as ProjectManifestV0_2_1Impl;
   }
 
   toDeployment(): string | undefined {