+++ packages/node/src/indexer/ds-processor.service.spec.ts
@@ -1,38 +1,38 @@
-// Copyright 2020-2021 OnFinality Limited authors & contributors
+// Copyright 2020-2022 OnFinality Limited authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
 import path from 'path';
-import { ProjectManifestVersioned, isCustomDs } from '@subql/common';
+import { isCustomDs } from '@subql/common';
 import { SubqlCustomDatasource } from '@subql/types';
-import { SubqueryProject } from '../configure/project.model';
+import { GraphQLSchema } from 'graphql';
+import { SubqueryProject } from '../configure/SubqueryProject';
 import { DsProcessorService } from './ds-processor.service';
 
-function getTestProject(extraDataSources?: SubqlCustomDatasource[]) {
-  return new SubqueryProject(
-    new ProjectManifestVersioned({
-      specVersion: '0.2.0',
-      version: '0.0.0',
-      network: {
-        genesisHash: '0x',
-        endpoint: 'wss://polkadot.api.onfinality.io/public-ws',
-      },
-      schema: './schema.graphql',
-      dataSources: [
-        {
-          kind: 'substrate/Jsonfy',
-          processor: { file: 'contract-processors/dist/jsonfy.js' },
-          startBlock: 1,
-          mapping: {
-            handlers: [
-              { handler: 'testSandbox', kind: 'substrate/JsonfyEvent' },
-            ],
-          },
+function getTestProject(
+  extraDataSources?: SubqlCustomDatasource[],
+): SubqueryProject {
+  return {
+    network: {
+      genesisHash: '0x',
+      endpoint: 'wss://polkadot.api.onfinality.io/public-ws',
+    },
+    dataSources: [
+      {
+        kind: 'substrate/Jsonfy',
+        processor: { file: 'contract-processors/dist/jsonfy.js' },
+        startBlock: 1,
+        mapping: {
+          entryScript: '',
+          handlers: [{ handler: 'testSandbox', kind: 'substrate/JsonfyEvent' }],
         },
-        ...extraDataSources,
-      ],
-    } as any),
-    path.resolve(__dirname, '../../../'),
-  );
+      },
+      ...extraDataSources,
+    ] as any,
+    id: 'test',
+    root: path.resolve(__dirname, '../../../'),
+    schema: new GraphQLSchema({}),
+    templates: [],
+  };
 }
 
 describe('DsProcessorService', () => {
@@ -45,8 +45,9 @@ describe('DsProcessorService', () => {
   });
 
   it('can validate custom ds', async () => {
-    await service.validateCustomDs();
-    await expect(service.validateCustomDs()).resolves.not.toThrow();
+    await expect(
+      service.validateProjectCustomDatasources(),
+    ).resolves.not.toThrow();
   });
 
   it('can catch an invalid datasource kind', async () => {
@@ -62,7 +63,7 @@ describe('DsProcessorService', () => {
     project = getTestProject([badDs]);
     service = new DsProcessorService(project);
 
-    await expect(service.validateCustomDs()).rejects.toThrow();
+    await expect(service.validateProjectCustomDatasources()).rejects.toThrow();
   });
 
   it('can run a custom ds processor', () => {