+++ packages/node/src/indexer/ds-processor.service.ts
@@ -1,4 +1,4 @@
-// Copyright 2020-2021 OnFinality Limited authors & contributors
+// Copyright 2020-2022 OnFinality Limited authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
 import fs from 'fs';
@@ -7,17 +7,19 @@ import { Injectable } from '@nestjs/common';
 import { isCustomDs } from '@subql/common';
 import {
   SubqlCustomDatasource,
+  SubqlDatasource,
   SubqlDatasourceProcessor,
   SubqlNetworkFilter,
 } from '@subql/types';
 import { VMScript } from '@subql/x-vm2';
-import { SubqueryProject } from '../configure/project.model';
+import { SubqueryProject } from '../configure/SubqueryProject';
 import { getLogger } from '../utils/logger';
 import { Sandbox } from './sandbox.service';
 
 export interface DsPluginSandboxOption {
   root: string;
   entry: string;
+  script: string;
 }
 
 const logger = getLogger('ds-sandbox');
@@ -49,8 +51,8 @@ export class DsProcessorService {
   } = {};
   constructor(private project: SubqueryProject) {}
 
-  async validateCustomDs(): Promise<void> {
-    for (const ds of this.project.dataSources.filter(isCustomDs)) {
+  async validateCustomDs(datasources: SubqlCustomDatasource[]): Promise<void> {
+    for (const ds of datasources) {
       const processor = this.getDsProcessor(ds);
       /* Standard validation applicable to all custom ds and processors */
       if (ds.kind !== processor.kind) {
@@ -80,6 +82,10 @@ export class DsProcessorService {
     }
   }
 
+  async validateProjectCustomDatasources(): Promise<void> {
+    await this.validateCustomDs((this.project.dataSources as SubqlDatasource[]).filter(isCustomDs));
+  }
+
   getDsProcessor<D extends string, T extends SubqlNetworkFilter>(
     ds: SubqlCustomDatasource<string, T>,
   ): SubqlDatasourceProcessor<D, T> {
@@ -88,8 +94,9 @@ export class DsProcessorService {
     }
     if (!this.processorCache[ds.processor.file]) {
       const sandbox = new DsPluginSandbox({
-        root: this.project.path,
+        root: this.project.root,
         entry: ds.processor.file,
+        script: null /* TODO get working with Readers, same as with sandbox */,
       });
       try {
         this.processorCache[ds.processor.file] = sandbox.getDsPlugin<D, T>();
@@ -118,7 +125,7 @@ export class DsProcessorService {
     for (const [name, { file }] of ds.assets) {
       // TODO update with https://github.com/subquery/subql/pull/511
       try {
-        res[name] = fs.readFileSync(path.join(this.project.path, file), {
+        res[name] = fs.readFileSync(file, {
           encoding: 'utf8',
         });
       } catch (e) {