+++ packages/node/src/indexer/fetch.service.spec.ts
@@ -1,12 +1,16 @@
-// Copyright 2020-2021 OnFinality Limited authors & contributors
+// Copyright 2020-2022 OnFinality Limited authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
 import path from 'path';
 import { EventEmitter2 } from '@nestjs/event-emitter';
-import { ProjectManifestVersioned } from '@subql/common';
-import { SubqlDatasourceKind, SubqlHandlerKind } from '@subql/types';
+import {
+  SubqlDatasourceKind,
+  SubqlHandlerKind,
+  SubqlMapping,
+} from '@subql/types';
+import { GraphQLSchema } from 'graphql';
 import { NodeConfig } from '../configure/NodeConfig';
-import { SubqueryProject } from '../configure/project.model';
+import { SubqueryProject } from '../configure/SubqueryProject';
 import { fetchBlocksBatches } from '../utils/substrate';
 import { ApiService } from './api.service';
 import { Dictionary, DictionaryService } from './dictionary.service';
@@ -91,6 +95,7 @@ const mockDictionaryRet: Dictionary = {
     indexerHealthy: true,
     indexerNodeVersion: '0.16.1',
     queryNodeVersion: '0.6.0',
+    rowCountEstimate: [{ table: '', estimate: 0 }],
   },
   //simulate after last process height update to 1000
   batchBlocks: [1000],
@@ -107,6 +112,7 @@ const mockDictionaryNoBatches: Dictionary = {
     indexerHealthy: true,
     indexerNodeVersion: '0.16.1',
     queryNodeVersion: '0.6.0',
+    rowCountEstimate: [{ table: '', estimate: 0 }],
   },
   batchBlocks: [],
 };
@@ -122,6 +128,7 @@ const mockDictionaryBatches: Dictionary = {
     indexerHealthy: true,
     indexerNodeVersion: '0.16.1',
     queryNodeVersion: '0.6.0',
+    rowCountEstimate: [{ table: '', estimate: 0 }],
   },
   batchBlocks: [14000, 14200, 14300, 14500, 14600, 14700, 14800, 14900],
 };
@@ -157,53 +164,53 @@ function mockDictionaryService3(): DictionaryService {
     getDictionary: jest.fn(() => mockDictionaryNoBatches),
   } as any;
 }
-
 function testSubqueryProject(): SubqueryProject {
-  const project = new SubqueryProject(
-    new ProjectManifestVersioned({
-      specVersion: '0.0.1',
-      network: {
-        endpoint: 'wss://polkadot.api.onfinality.io/public-ws',
-        types: {
-          TestType: 'u32',
-        },
+  return {
+    network: {
+      endpoint: 'wss://polkadot.api.onfinality.io/public-ws',
+    },
+    chainTypes: {
+      types: {
+        TestType: 'u32',
       },
-      dataSources: [],
-    } as any),
-    '',
-  );
-  return project;
+    },
+    dataSources: [],
+    id: 'test',
+    root: './',
+    schema: new GraphQLSchema({}),
+    templates: [],
+  };
 }
 
 function testSubqueryProjectV0_2_0(): SubqueryProject {
-  const project = new SubqueryProject(
-    new ProjectManifestVersioned({
-      specVersion: '0.2.0',
-      network: {
-        genesisHash: '0x',
-        endpoint: 'wss://polkadot.api.onfinality.io/public-ws',
-      },
-      dataSources: [
-        {
-          kind: 'substrate/Jsonfy',
-          processor: {
-            file: 'contract-processors/dist/jsonfy.js',
-          },
-          startBlock: 1,
-          mapping: {
-            handlers: [
-              {
-                handler: 'handleEvent',
-                kind: 'substrate/JsonfyEvent',
-              },
-            ],
-          },
+  return {
+    network: {
+      genesisHash: '0x',
+      dictionary: `https://api.subquery.network/sq/subquery/dictionary-polkadot`,
+    },
+    dataSources: [
+      {
+        kind: 'substrate/Jsonfy',
+        processor: {
+          file: 'contract-processors/dist/jsonfy.js',
         },
-      ],
-    } as any),
-    path.resolve(__dirname, '../../../'),
-  );
-  return project;
+        startBlock: 1,
+        mapping: {
+          entryScript: '',
+          handlers: [
+            {
+              handler: 'handleEvent',
+              kind: 'substrate/JsonfyEvent',
+            },
+          ],
+        },
+      },
+    ] as any,
+    id: 'test',
+    schema: new GraphQLSchema({}),
+    root: path.resolve(__dirname, '../../../'),
+    templates: [],
+  };
 }
 
 function createFetchService(
@@ -275,6 +282,11 @@ describe('FetchService', () => {
 
   it('loop until shutdown', async () => {
     const batchSize = 20;
+    (fetchBlocksBatches as jest.Mock).mockImplementation((api, blockArray) =>
+      blockArray.map((height) => ({
+        block: { block: { header: { number: { toNumber: () => height } } } },
+      })),
+    );
     const dictionaryService = new DictionaryService(project);
 
     const fetchService = createFetchService(
@@ -297,14 +309,15 @@ describe('FetchService', () => {
 
   it("skip use dictionary once if dictionary 's lastProcessedHeight < startBlockHeight ", async () => {
     const batchSize = 20;
-    project.projectManifest.asV0_0_1.network.dictionary =
+    project.network.dictionary =
       'https://api.subquery.network/sq/subquery/dictionary-polkadot';
-    project.projectManifest.asV0_0_1.dataSources = [
+    project.dataSources = [
       {
         name: 'runtime',
         kind: SubqlDatasourceKind.Runtime,
         startBlock: 1,
         mapping: {
+          entryScript: '',
           handlers: [
             {
               handler: 'handleCall',
@@ -358,77 +371,17 @@ describe('FetchService', () => {
     expect((fetchService as any).useDictionary).toBeTruthy();
   }, 500000);
 
-  it('skip use dictionary once if getDictionary(api failure) return undefined ', async () => {
-    const batchSize = 20;
-    project.projectManifest.asV0_0_1.network.dictionary =
-      'https://api.subquery.network/sq/subquery/dictionary-polkadot';
-    project.projectManifest.asV0_0_1.dataSources = [
-      {
-        name: 'runtime',
-        kind: SubqlDatasourceKind.Runtime,
-        startBlock: 1,
-        mapping: {
-          handlers: [
-            {
-              handler: 'handleBond',
-              kind: SubqlHandlerKind.Event,
-              filter: {
-                module: 'staking',
-                method: 'Bonded',
-              },
-            },
-          ],
-        },
-      },
-    ];
-    const dictionaryService = mockDictionaryService2();
-    const dsPluginService = new DsProcessorService(project);
-    const eventEmitter = new EventEmitter2();
-    const fetchService = new FetchService(
-      apiService,
-      new NodeConfig({ subquery: '', subqueryName: '', batchSize }),
-      project,
-      dictionaryService,
-      dsPluginService,
-      eventEmitter,
-    );
-    const nextEndBlockHeightSpy = jest.spyOn(
-      fetchService as any,
-      `nextEndBlockHeight`,
-    );
-    const dictionaryValidationSpy = jest.spyOn(
-      fetchService as any,
-      `dictionaryValidation`,
-    );
-    await fetchService.init();
-    (fetchService as any).latestFinalizedHeight = 1005;
-    (fetchService as any).latestBufferedHeight = undefined;
-    (fetchService as any).latestProcessedHeight = undefined;
-    const loopPromise = fetchService.startLoop(1000);
-    eventEmitter.on(`blocknumber_queue_size`, (nextBufferSize) => {
-      // [1000,1001,1002,1003,1004]
-      if (nextBufferSize.value >= 5) {
-        fetchService.onApplicationShutdown();
-      }
-    });
-    await loopPromise;
-    //validation will not be called as dictionary is undefined
-    expect(dictionaryValidationSpy).toHaveBeenCalledTimes(0);
-    expect(nextEndBlockHeightSpy).toHaveBeenCalledTimes(1);
-    //we expect after use the original method, next loop will still use dictionary by default
-    expect((fetchService as any).useDictionary).toBeTruthy();
-  }, 500000);
-
   it('set last buffered Height to dictionary last processed height when dictionary returned batch is empty, and then start use original method', async () => {
     const batchSize = 20;
-    project.projectManifest.asV0_0_1.network.dictionary =
+    project.network.dictionary =
       'https://api.subquery.network/sq/subquery/dictionary-polkadot';
-    project.projectManifest.asV0_0_1.dataSources = [
+    project.dataSources = [
       {
         name: 'runtime',
         kind: SubqlDatasourceKind.Runtime,
         startBlock: 1,
         mapping: {
+          entryScript: '',
           handlers: [
             {
               handler: 'handleBond',
@@ -476,14 +429,15 @@ describe('FetchService', () => {
 
   it('fill the dictionary returned batches to nextBlockBuffer', async () => {
     const batchSize = 20;
-    project.projectManifest.asV0_0_1.network.dictionary =
+    project.network.dictionary =
       'https://api.subquery.network/sq/subquery/dictionary-polkadot';
-    project.projectManifest.asV0_0_1.dataSources = [
+    project.dataSources = [
       {
         name: 'runtime',
         kind: SubqlDatasourceKind.Runtime,
         startBlock: 1,
         mapping: {
+          entryScript: '',
           handlers: [
             {
               handler: 'handleBond',