+++ packages/node/src/indexer/poi.service.ts
@@ -1,11 +1,11 @@
-// Copyright 2020-2021 OnFinality Limited authors & contributors
+// Copyright 2020-2022 OnFinality Limited authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
 import { Injectable, OnApplicationShutdown } from '@nestjs/common';
 import { hexToU8a } from '@polkadot/util';
 import { Sequelize } from 'sequelize';
 import { NodeConfig } from '../configure/NodeConfig';
-import { SubqueryProject } from '../configure/project.model';
+import { SubqueryProject } from '../configure/SubqueryProject';
 import { PoiFactory, PoiRepo, ProofOfIndex } from './entities/Poi.entity';
 
 const DEFAULT_PARENT_HASH = hexToU8a('0x00');