+++ packages/node/src/yargs.ts
@@ -1,4 +1,4 @@
-// Copyright 2020-2021 OnFinality Limited authors & contributors
+// Copyright 2020-2022 OnFinality Limited authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
 import { hideBin } from 'yargs/helpers';
@@ -10,7 +10,7 @@ export function getYargsOption() {
     subquery: {
       alias: 'f',
       demandOption: false,
-      describe: 'Local path of the subquery project',
+      describe: 'Local path or IPFS cid of the subquery project',
       type: 'string',
     },
     'subquery-name': {
@@ -26,6 +26,7 @@ export function getYargsOption() {
       type: 'string',
     },
     local: {
+      deprecated: true,
       type: 'boolean',
       demandOption: false,
       describe: 'Use local mode',
@@ -121,12 +122,16 @@ export function getYargsOption() {
       type: 'boolean',
       default: false,
     },
+    ipfs: {
+      demandOption: false,
+      describe: 'IPFS gateway endpoint',
+      type: 'string',
+    },
     port: {
       alias: 'p',
       demandOption: false,
       describe: 'The port the service will bind to',
       type: 'number',
-      default: 3000,
     },
   });
 }