+++ packages/validator/src/validator.spec.ts
@@ -1,4 +1,4 @@
-// Copyright 2020-2021 OnFinality Limited authors & contributors
+// Copyright 2020-2022 OnFinality Limited authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
 import {commonRules} from './rules';
@@ -7,16 +7,15 @@ import {Validator} from './validator';
 describe('Validator', () => {
   let v: Validator;
 
-  beforeAll(() => {
+  beforeAll(async () => {
     const url = 'https://github.com/subquery/tutorials-block-timestamp';
-    v = new Validator(url);
+    v = await Validator.create(url);
     v.addRule(...commonRules);
   });
 
   it('should validate get reports', async () => {
     const result = await v.getValidateReports();
-    expect(result.length).toBe(8);
-    expect(result.filter((r) => r.valid).length).toBe(8);
+    expect(result.filter((r) => r.valid).length).toBe(result.length);
   });
 
   it('should return validate result', async () => {