+++ packages/cli/CHANGELOG.md
@@ -8,6 +8,26 @@ All logs must start with the format: [x.y.z] - yyyy-mm-dd
 
 ## [Unreleased]
 
+## [0.19.0] - 2022-01-21
+### Added
+Allow use `subql build` self contained js as instructed by `exports` fields under`package.json` (#753)
+### Changed
+- Remove .github directory from new projects (#763)
+
+## [0.18.0] - 2022-01-18
+### Added
+- Filter project templates by selected specVersion (#751)
+
+## [0.17.0] - 2022-01-13
+### Added
+- Warning and advice for use manifest specVersion v0.0.1 (#730)
+- Message for install git suggestion (#733)
+- Allow user to select templates when run `subql init` (#712)
+
+## [0.16.2] - 2021-12-16
+### Changed
+- Update subql/validator dependencies (#720)
+
 ## [0.16.1] - 2021-11-30
 ### Fixed
 - Remove auto generated enum types and interfaces (#680)
@@ -118,7 +138,11 @@ All logs must start with the format: [x.y.z] - yyyy-mm-dd
 - support subcommand codegen
 - support subcommand init
 
-[Unreleased]: https://github.com/subquery/subql/compare/cli/0.16.1...HEAD
+[Unreleased]: https://github.com/subquery/subql/compare/cli/0.19.0...HEAD
+[0.19.0]: https://github.com/subquery/subql/compare/cli/0.18.0...cli/0.19.0
+[0.18.0]: https://github.com/subquery/subql/compare/cli/0.17.0...cli/0.18.0
+[0.17.0]: https://github.com/subquery/subql/compare/cli/0.16.2...cli/0.17.0
+[0.16.2]: https://github.com/subquery/subql/compare/cli/0.16.1...cli/0.16.2
 [0.16.1]: https://github.com/subquery/subql/compare/cli/0.16.0...cli/0.16.1
 [0.16.0]: https://github.com/subquery/subql/compare/cli/0.15.0...cli/0.16.0
 [0.15.0]: https://github.com/subquery/subql/compare/cli/0.14.0...cli/0.15.0